WIP: syncing between devices

+++ backend/src/main/java/io/github/filipolszewski/cookbook/dto/recipeingredient/RecipeIngredientAddRequest.java
@@ -3,7 +3,7 @@ package io.github.filipolszewski.cookbook.dto.recipeingredient;
 import jakarta.validation.constraints.NotNull;
 
 public record RecipeIngredientAddRequest(
-    @NotNull Integer ingredientId,
+    @NotNull Long ingredientId,
     @NotNull Double amount,
     @NotNull String unit
 ) {}
